parse fn files as well
